# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateEmployeeDto": {
        "properties": {
          "departmentId": {
            "enum": [
              "engineering",
              "marketing",
              "sales",
              "hr"
            ],
            "example": "engineering",
            "type": "string"
          },
          "isActive": {
            "description": "true = Active, false = In Active",
            "example": true,
            "type": "boolean"
          },
          "joinDate": {
            "example": "2026-09-01",
            "format": "date",
            "type": "string"
          },
          "name": {
            "description": "Trimmed and NFC-normalized; 1–100 code points.",
            "example": "Dana Lee",
            "maxLength": 100,
            "type": "string"
          },
          "salary": {
            "description": "Decimal string, 0–9999999999.99",
            "example": "62000.00",
            "pattern": "^\\d{1,10}(\\.\\d{1,2})?$",
            "type": "string"
          }
        },
        "required": [
          "name",
          "departmentId",
          "salary",
          "joinDate",
          "isActive"
        ],
        "type": "object"
      },
...
      "EmployeeDto": {
        "properties": {
          "departmentId": {
            "enum": [
              "engineering",
              "marketing",
              "sales",
              "hr"
            ],
            "type": "string"
          },
          "departmentName": {
            "example": "Engineering",
            "type": "string"
          },
          "id": {
            "example": 106,
            "type": "number"
          },
          "isActive": {
            "example": true,
            "type": "boolean"
          },
          "joinDate": {
            "example": "2026-09-01",
            "format": "date",
            "type": "string"
          },
          "lastUpdatedDate": {
            "example": "2026-10-01",
            "format": "date",
            "type": "string"
          },
          "name": {
            "example": "Dana Lee",
            "type": "string"
          },
          "salary": {
            "description": "Present for Admin only; the key is omitted for Viewer.",
            "example": "62000.00",
            "type": "string"
          },
          "version": {
            "example": 1,
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "departmentId",
          "departmentName",
          "joinDate",
          "isActive",
          "lastUpdatedDate",
          "version"
        ],
        "type": "object"
      },
      "EmployeeListMetaDto": {
        "properties": {
          "page": {
            "example": 1,
            "type": "number"
          },
          "pageSize": {
            "example": 20,
            "type": "number"
          },
          "requestId": {
            "format": "uuid",
            "type": "string"
          },
          "sortBy": {
            "enum": [
              "id",
              "name",
              "department",
              "joinDate",
              "isActive",
              "lastUpdatedDate",
              "salary"
            ],
            "type": "string"
          },
          "sortOrder": {
            "enum": [
              "asc",
              "desc"
            ],
            "type": "string"
          },
          "total": {
            "example": 5,
            "type": "number"
          },
          "totalPages": {
            "description": "0 when total is 0",
            "example": 1,
            "type": "number"
          }
        },
        "required": [
          "requestId",
          "page",
          "pageSize",
          "total",
          "totalPages",
          "sortBy",
          "sortOrder"
        ],
        "type": "object"
      },
...
      "ErrorBodyDto": {
        "properties": {
          "code": {
            "example": "VALIDATION_ERROR",
            "type": "string"
          },
          "currentReportId": {
            "description": "REPORT_IN_PROGRESS only",
            "format": "uuid",
            "nullable": true,
            "type": "object"
          },
          "currentVersion": {
            "description": "VERSION_CONFLICT only",
            "type": "number"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/FieldErrorDto"
            },
            "type": "array"
          },
          "message": {
            "example": "Please correct the highlighted fields.",
            "type": "string"
          },
          "requestId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "type": "object"
      },
      "ErrorEnvelopeDto": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBodyDto"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "FieldErrorDto": {
        "properties": {
          "code": {
            "example": "DECIMAL_SCALE_EXCEEDED",
            "type": "string"
          },
          "field": {
            "example": "salary",
            "type": "string"
          },
          "message": {
            "example": "Salary must have at most 2 decimal places.",
            "type": "string"
          }
        },
        "required": [
          "field",
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "MetaDto": {
        "properties": {
          "requestId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "requestId"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "cookie": {
        "in": "cookie",
        "name": "employee_console.local.sid",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "NestJS API for Employee Console. Browser clients authenticate with the session cookie (Google OIDC) and send X-CSRF-Token on mutations. /internal/v1 is for the n8n worker/scheduler (bearer tokens).",
    "title": "Employee Console API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/employees": {
      "get": {
        "operationId": "Employees_list",
        "parameters": [
          {
            "description": "Case-insensitive name contains; trimmed; ≤100 characters",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "departmentId",
            "required": false,
            "schema": {
              "enum": [
                "engineering",
                "marketing",
                "sales",
                "hr"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "all",
              "enum": [
                "all",
                "active",
                "inactive"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 1000000,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "salary is Admin-only",
            "in": "query",
            "name": "sortBy",
            "required": false,
            "schema": {
              "default": "id",
              "enum": [
                "id",
                "name",
                "department",
                "joinDate",
                "isActive",
                "lastUpdatedDate",
                "salary"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sortOrder",
            "required": false,
            "schema": {
              "default": "asc",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "items": {
                        "$ref": "#/components/schemas/EmployeeDto"
                      },
                      "type": "array"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/EmployeeListMetaDto"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Viewer responses omit the salary key",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "VALIDATION_ERROR / INVALID_QUERY / INVALID_IDEMPOTENCY_KEY / MALFORMED_JSON",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "UNAUTHENTICATED / SESSION_EXPIRED / INVALID_SERVICE_TOKEN",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "FORBIDDEN / ACCOUNT_NOT_ALLOWED / CSRF_INVALID",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "RATE_LIMITED / REPORT_QUOTA_EXCEEDED",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "INTERNAL_ERROR",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "employees"
        ]
      },
      "post": {
        "operationId": "Employees_create",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "UUID generated once per create intent",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmployeeDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/EmployeeDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Success",
            "headers": {
              "ETag": {
                "description": "Quoted employee version, e.g. \"1\" — send it back as If-Match",
                "schema": {
                  "type": "string"
                }
              },
              "Idempotency-Replayed": {
                "description": "\"true\" when the response was replayed for a repeated Idempotency-Key",
                "schema": {
                  "type": "string"
                }
              },
              "Location": {
                "description": "Path of the created resource",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "VALIDATION_ERROR / INVALID_QUERY / INVALID_IDEMPOTENCY_KEY / MALFORMED_JSON",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "UNAUTHENTICATED / SESSION_EXPIRED / INVALID_SERVICE_TOKEN",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "FORBIDDEN / ACCOUNT_NOT_ALLOWED / CSRF_INVALID",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "VERSION_CONFLICT / IDEMPOTENCY_CONFLICT / REPORT_IN_PROGRESS / STALE_LEASE / REPORT_ALREADY_COMPLETED",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "PAYLOAD_TOO_LARGE",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "UNSUPPORTED_MEDIA_TYPE",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "RATE_LIMITED / REPORT_QUOTA_EXCEEDED",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "INTERNAL_ERROR",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "employees"
        ]
      }
    },
...
  },
  "servers": [],
...
}